Increase the main/watchdog period even more
Arguably the watchdog is kind of useless in this project since so few
things can go wrong.

As a compromise we increase the feed time to 3 minutes.

Also now the watchdog timeout is not a multiple of the main period anymore
so we basically have only a single chance to feed it.